Extracted template from top-rated-films component
