+++ src/components/top-rated-films/template.js
@@ -1,0 +1,13 @@
+/**
+ * Creates top rated films template
+ * @return {String} template
+ */
+const template = () => {
+  return `<section class="films-list--extra films-list--top-rated">
+          <h2 class="films-list__title">Top rated</h2>
+          <div class="films-list__container">
+          </div>
+        </section>`;
+};
+
+export {template};
